requirements.txt
pyrogram==2.0.106
tgcrypto
websockets>=12.0
python-dotenv>=1.0
